+++ analysis/azimuthal/basic/calc/azim_stream2_calc.py
@@ -27,7 +27,7 @@ def process_t(t):
     # データの形状から半径方向のビン数を取得
     nr = rho.shape[1]
     R = (np.arange(nr) + 0.5) * config.dx
-    phi = np.zeros_like(rho)
+    phi = np.zeros_like(rho, dtype=np.float32)
 
     # ベクトル化版：r方向の積分（z=0の行）
     # 従来版: for r in range(1, nr): phi[0, r] = phi[0, r - 1] + ...
