+++ analysis/azimuthal/eliassen/calc/azim_N2_calc.py
@@ -35,7 +35,7 @@ g = 9.80665
 
 def process_t(t):
     b = np.load(os.path.join(config.get_data_path('azim', 'eliassen', 'buoyancy'), f"t{str(t).zfill(3)}.npy"))
-    db_dz = np.zeros((config.nz - 1, nr))
+    db_dz = np.zeros((config.nz - 1, nr), dtype=np.float32)
     for z in range(config.nz - 1):
         db_dz[z, :] = (b[z + 1, :] - b[z, :]) / (vgrid[z + 1] - vgrid[z])
     np.save(os.path.join(output_folder, f"t{str(t).zfill(3)}.npy"), db_dz)
